listening_post/src/requirements.txt
Flask==3.0.3
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
mysql_connector_repackaged==0.3.1
